requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai>=2.0.0
pillow==10.2.0
python-dotenv==1.0.1
aiofiles==23.2.1
google-generativeai>=0.8.0
libsql-client>=0.3.0
